User's Guide

User's Guide

Kernel

Kernel can be specified as a matrix or as a function. Initialize a kernel as a function:

getOpticsKernels — Method.

getOpticsKernels(name::String, alpha::Float64 = 1.)

name - name of a kernel,

alpha - kernel function parameter.

Returns function of 2 variables.

Available kernels:

rectangular:

\[K(x, y) = \begin{cases} 1, \text{if} |x-y|/\alpha < 1 \\ 0, \text{otherwise} \end{cases}\]

diffraction:

\[K(x, y) = \left(\frac{sin(\pi (x-y) /s_0)}{\pi (x-y) /s_0}\right)^2\]

\[s_0=\alpha/0.886\]

gaussian:

\[K(x, y) = \frac{2}{\alpha}\sqrt{\frac{\ln2}{\pi}}e^{4\ln2\left(\frac{x-y}{\alpha}\right)^2}\]

triangular:

\[K(x, y) = \begin{cases} \left(1 - \frac{|x-y|}{\alpha}\right)/\alpha, \text{if} |x-y|/\alpha < 1 \\ 0, \text{otherwise} \end{cases}\]

dispersive:

\[K(x, y) = \frac{\alpha}{2 \pi}\left((x-y)^2 + \left(\frac{\alpha}{2}\right)^2\right)\]

exponential:

\[K(x, y) = \frac{\ln2}{\alpha}e^{2\ln2\frac{|x-y|}{\alpha}}\]

heaviside:

\[K(x, y) = \begin{cases} 1, \text{if} x>0 \\ 0, \text{otherwise} \end{cases}\]

discretize_kernel — Method.

    discretize_kernel(basis::Basis, kernel::Function, xs::Array{Float64, 1})

Returns matrix n*m, where n - number of basis function, m - number of knots.

Basis

Basis — Type.

Abstract type for all bases.

BaseFunction — Type.

BaseFunction(f, support::Tuple{Float64,Float64})
BaseFunction(f, a::Float64, b::Float64)

Class for function with its support.

omega — Method.

omega(basis::Basis, deg::Int64)

Returns matrix of the mean values of derivatives of order deg

Fourier basis

FourierBasis — Type.

FourierBasis(a::Float64, b::Float64, n::Int64)

a, b – the beginning and the end of the segment

n – number of basis functions

Cubic Spline basis

CubicSplineBasis — Type.

CubicSplineBasis(
    knots::Array{Float64},
    boundary_condition::Union{Tuple{Union{String, Nothing}, Union{String, Nothing}}, Nothing, String}=nothing
    )

knots – knots of spline,

boundary_condition – boundary conditions of basis functions. If tuple, the first element affects left bound, the second element affects right bound. If string, both sides are affected. Possible options: "dirichlet", nothing.

Legendre polynomial basis

LegendreBasis — Type.

LegendreBasis(a::Float64, b::Float64, n::Int64)

a, b – the beginning and the end of the segment

n – number of basis functions

Bernstein polynomial basis

BernsteinBasis — Type.

BernsteinBasis(
    a::Float64, b::Float64, n::Int64,
    boundary_condition::Union{Tuple{Union{String, Nothing}, Union{String, Nothing}}, Nothing, String}=nothing
    )

a, b – the beginning and the end of the segment

n – number of basis functions

boundary_condition – boundary conditions of basis functions. If tuple, the first element affects left bound, the second element affects right bound. If string, both sides are affected. Possible options: "dirichlet", nothing.

Model

Matrix unfolder

GaussErrorMatrixUnfolder — Type.

GaussErrorMatrixUnfolder(
    omegas::Array{Array{Float64, 2} ,1},
    method::String="EmpiricalBayes",
    alphas::Union{Array{Float64, 1}, Nothing}=nothing,
    )

omegas – array of matrices that provide information about basis functions

method – constant selection method, possible options: "EmpiricalBayes" and "User"

alphas – array of constants, in case method="User" should be provided by user

Solve model for dicsrete data.

Unfolder

GaussErrorUnfolder — Type.

GaussErrorUnfolder(
    basis::Basis,
    omegas::Array,
    method::String="EmpiricalBayes",
    alphas::Union{Array{Float64, 1}, Nothing}=nothing,
    )

basis – basis for reconstruction

omegas – array of matrices that provide information about basis functions

method – constant selection method, possible options: "EmpiricalBayes" and "User"

alphas – array of constants, in case method="User" should be provided by user

Reconstruction

Matrix unfolder

solve — Method.

solve(
    unfolder::GaussErrorMatrixUnfolder,
    kernel::Array{Float64, 2},
    data::Array{Float64, 1},
    data_errors::Union{Array{Float64, 1}, Array{Float64, 2}},
    )

Returns dictionary with coefficients, errors and optimal constants.

Unfolder

solve — Method.

solve(
    gausserrorunfolder::GaussErrorUnfolder,
    kernel::Union{Function, Array{Float64, 2}},
    data::Union{Function, Array{Float64, 1}},
    data_errors::Union{Function, Array{Float64, 1}},
    y::Union{Array{Float64, 1}, Nothing},
    )

Returns dictionary with coefficients, errors and optimal constants.

Result

Missing docstring.

Missing docstring for PhiVec(coeff::Array{Float64}, basis::Basis, sig::Array{Float64}). Check Documenter's build log for details.

call — Method.

call(phivec::PhiVec, x::Float64)

Returns solution function value in given point x.

errors — Method.

errors(phi::PhiVec, x::Float64)

Returns error in given point x.

errors — Method.

errors(phi::PhiVec, xs::Array{Float64})

Returns error in given point set xs.