User's Guide

User's Guide

Kernel

Kernel can be specified as a matrix or as a function. It is possible to set arbitrary function of 2 variables or use one of predefined kernels.

Main.StatReg.getOpticsKernels — Method.

getOpticsKernels(name::String, alpha::Real = 1.)

Arguments

name - name of a kernel

alpha - kernel function parameter

Returns: kernel, function of 2 variables.

Available kernels:

rectangular:

\[K(x, y) = \begin{cases} 1, \text{if } \frac{|x-y|}{\alpha} < 1 \ \text{; } 0 \text{ otherwise} \end{cases}\]

diffraction:

\[K(x, y) = \left(\frac{sin(\frac{\pi (x-y)}{s_0})}{\frac{\pi (x-y)}{s_0}}\right)^2\]

\[s_0 = \frac{\alpha}{0.886}\]

gaussian:

\[K(x, y) = \frac{2}{\alpha}\sqrt{\frac{\ln2}{\pi}}e^{4\ln2\left(\frac{x-y}{\alpha}\right)^2}\]

triangular:

\[K(x, y) = \begin{cases} \frac{1 - \frac{|x-y|}{\alpha}}{\alpha}, \text{if } \frac{|x-y|}{\alpha} < 1 \ \text{; } 0 \text{ otherwise} \end{cases}\]

dispersive:

\[K(x, y) = \frac{\alpha}{2 \pi}\left((x-y)^2 + \left(\frac{\alpha}{2}\right)^2\right)\]

exponential:

\[K(x, y) = \frac{\ln2}{\alpha}e^{2\ln2\frac{|x-y|}{\alpha}}\]

heaviside:

\[K(x, y) = \begin{cases} 1, \text{if } x>0 \ \text{; } 0 \text{ otherwise} \end{cases}\]

source

Main.StatReg.discretize_kernel — Method.

discretize_kernel(basis::Basis, kernel::Function, data_points::AbstractVector{<:Real})

Arguments

basis – basis

kernel – kernel function

data_points – array of data points

Returns: discretized kernel K::Array{Real, 2}, $K_{mn} = \int\limits_a^b K(x, y_n) \psi_m(x) dx$ - matrix of size n$\times$m, where $m$ - number of basis functions, $n$ - number of data points.

source

Basis

Main.StatReg.Basis — Type.

Abstract type for all bases.

source

Main.StatReg.BaseFunction — Type.

Type for function with its support.

BaseFunction(f, support::Tuple{<:Real,<:Real})
BaseFunction(f, a::Real, b::Real)

Fields

f – function (type depends on the basis)

support::Tuple{<:Real,<:Real} – support of the function

source

Main.StatReg.omega — Method.

omega(basis::Basis, ord::Int)

Arguments

basis - basis

ord - order of derivatives

Returns: Omega::Array{Real, 2}, $\Omega_{mn} = \int\limits_a^b \frac{d^{ord} \psi_m}{dx^{ord}} \frac{d^{ord} \psi_n}{dx^{ord}}$ - matrix of size n$\times$n of the mean values of derivatives of order ord, where n - number of functions in basis.

source

Fourier basis

Main.StatReg.FourierBasis — Type.

Fourier basis with length $2n+1$: {$0.5$, $\sin(\frac{\pi (x - \frac{a+b}{2})}{b-a})$, $\cos(\frac{\pi (x - \frac{a+b}{2})}{b-a})$, ..., $\sin(\frac{\pi n (x - \frac{a+b}{2})}{b-a})$, $\cos(\frac{\pi n (x - \frac{a+b}{2})}{b-a})$}.

FourierBasis(a::Real, b::Real, n::Int)

a, b – the beginning and the end of the segment n – number of basis functions

Fields

a::Real – beginning of the support

b::Real – end of the support

n::Int – number of basis functions

basis_functions::AbstractVector{BaseFunction} – array of basis functions

source

Cubic Spline basis

Main.StatReg.CubicSplineBasis — Type.

Cubic spline basis on given knots with length $n$, where n – length of knots array.

CubicSplineBasis(
    knots::AbstractVector{<:Real},
    boundary_condition::Union{Tuple{Union{String, Nothing}, Union{String, Nothing}}, Nothing, String}=nothing
    )

CubicSplineBasis(
    a::Real, b::Real, n::Int,
    boundary_condition::Union{Tuple{Union{String, Nothing}, Union{String, Nothing}}, Nothing, String}=nothing
    )

knots – knots of spline boundary_condition – boundary conditions of basis functions. If tuple, the first element affects left bound, the second element affects right bound. If string, both sides are affected. Possible options: "dirichlet", nothing

Fields

a::Real – beginning of the support, matches the first element of the array knots

b::Real – end of the support, matches the last element of the array knots

knots::AbstractVector{<:Real} – array of points on which the spline is built

basis_functions::AbstractVector{BaseFunction} – array of basis functions

source

Legendre polynomials basis

Main.StatReg.LegendreBasis — Type.

Legendre polynomials basis with length $n$.

LegendreBasis(a::Real, b::Real, n::Int)

a, b – the beginning and the end of the support n – number of basis functions

Fields

a::Real – beginning of the support

b::Real – end of the support

basis_functions::AbstractVector{BaseFunction} – array of basis functions

source

Bernstein polynomials basis

Main.StatReg.BernsteinBasis — Type.

Bernstein polynomials basis.

BernsteinBasis(
    a::Real, b::Real, n::Int,
    boundary_condition::Union{Tuple{Union{String, Nothing}, Union{String, Nothing}}, Nothing, String}=nothing
    )

a, b – the beginning and the end of the segment n – number of basis functions boundary_condition – boundary conditions of basis functions. If tuple, the first element affects left bound, the second element affects right bound. If string, both sides are affected. Possible options: "dirichlet", nothing.

Fields

a::Real – beginning of the support

b::Real – end of the support

basis_functions::AbstractVector{BaseFunction} – array of basis functions

boundary_condition::Tuple{Union{String, Nothing}, Union{String, Nothing}} – boundary conditions of basis functions. If tuple, the first element affects left bound, the second element affects right bound. If string, both sides are affected. Possible options: "dirichlet", nothing.

source

Gaussian noise distribution

Model

Main.StatReg.GaussErrorMatrixUnfolder — Type.

Model for discrete data and kernel.

GaussErrorMatrixUnfolder(
    omegas::Array{Array{T, 2}, 1} where T<:Real,
    method::String="EmpiricalBayes";
    alphas::Union{AbstractVector{<:Real}, Nothing}=nothing,
    lower::Union{AbstractVector{<:Real}, Nothing}=nothing,
    higher::Union{AbstractVector{<:Real}, Nothing}=nothing,
    initial::Union{AbstractVector{<:Real}, Nothing}=nothing
    )

omegas – array of matrices that provide information about basis functions

method – constant selection method, possible options: "EmpiricalBayes" and "User"

alphas – array of constants, in case method="User" should be provided by user

lower – lowerer limits for alphas

higher – higherer limits for alphas

initial – unitial values for alphas

Fields

omegas::Array{Array{T, 2}, 1} where T<:Real

n::Int – size of square omega matrix

method::String

alphas::Union{AbstractVector{<:Real}, Nothing}

lower::Union{AbstractVector{<:Real}, Nothing}

higher::Union{AbstractVector{<:Real}, Nothing}

initial::Union{AbstractVector{<:Real}, Nothing}

source

Main.StatReg.GaussErrorUnfolder — Type.

Model for continuous kernel. Data can be either discrete or continuous.

GaussErrorUnfolder(
    basis::Basis,
    omegas::Array{Array{T, 2}, 1} where T<:Real,
    method::String="EmpiricalBayes";
    alphas::Union{AbstractVector{<:Real}, Nothing} =nothing,
    lower::Union{AbstractVector{<:Real}, Nothing} =nothing,
    higher::Union{AbstractVector{<:Real}, Nothing} =nothing,
    initial::Union{AbstractVector{<:Real}, Nothing} =nothing
    )

basis – basis for reconstruction

omegas – array of matrices that provide information about basis functions

method – constant selection method, possible options: "EmpiricalBayes" and "User"

alphas – array of constants, in case method="User" should be provided by user

lower – lowerer limits for alphas

higher – higherer limits for alphas

initial – unitial values for alphas

Fields

basis::Basis

solver::GaussErrorMatrixUnfolder

source

Reconstruction

Main.StatReg.solve — Method.

solve(
    unfolder::GaussErrorMatrixUnfolder,
    kernel::AbstractMatrix{<:Real},
    data::AbstractVector{<:Real},
    data_errors::AbstractVecOrMat{<:Real}
    )

Arguments

unfolder – model

kernel – discrete kernel

data – function values

data_errors – function errors

Returns: Dict{String, AbstractVector{Real}} with coefficients ("coeff"), errors ("errors") and optimal constants ("alphas").

source

Main.StatReg.solve — Function.

solve(
    unfolder::GaussErrorUnfolder,
    kernel::Union{Function, AbstractMatrix{<:Real}},
    data::Union{Function, AbstractVector{<:Real}},
    data_errors::Union{Function, AbstractVector{<:Real}},
    y::Union{AbstractVector{<:Real}, Nothing}=nothing,
    )

Arguments

unfolder – model

kernel – discrete or continuous kernel

data – function values

data_errors – function errors

y – points to calculate function values and its errors (when data is given as a function)

Returns: `Dict{String, AbstractVector{Real} with coefficients ("coeff"), errors ("errors") and optimal constants ("alphas").

source

Non-Gaussian noise distribution

Model

Main.StatReg.MCMCMatrixUnfolder — Type.

MCMC model for discrete data and kernel.

MCMCMatrixUnfolder(
    omegas::Array{Array{T, 2}, 1} where T<:Real,
    method::String="EmpiricalBayes";
    alphas::Union{AbstractVector{<:Real}, Nothing}=nothing,
    lower::Union{AbstractVector{<:Real}, Nothing}=nothing,
    higher::Union{AbstractVector{<:Real}, Nothing}=nothing,
    initial::Union{AbstractVector{<:Real}, Nothing}=nothing
    )

omegas – array of matrices that provide information about basis functions

method – constant selection method, possible options: "EmpiricalBayes" and "User"

alphas – array of constants, in case method="User" should be provided by user

lower – lowerer limits for alphas

higher – higherer limits for alphas

initial – unitial values for alphas

Fields

omegas::Array{Array{T, 2}, 1} where T<:Real

n::Int – size of square omega matrix

method::String

alphas::Union{AbstractVector{<:Real}, Nothing}

lower::Union{AbstractVector{<:Real}, Nothing}

higher::Union{AbstractVector{<:Real}, Nothing}

initial::Union{AbstractVector{<:Real}, Nothing}

source

Main.StatReg.MCMCUnfolder — Type.

MCMC model for continuous kernel. Data can be either discrete or continuous.

MCMCUnfolder(
    basis::Basis,
    omegas::Array{Array{T, 2}, 1} where T<:Real,
    method::String="EmpiricalBayes";
    alphas::Union{AbstractVector{<:Real}, Nothing}=nothing,
    lower::Union{AbstractVector{<:Real}, Nothing}=nothing,
    higher::Union{AbstractVector{<:Real}, Nothing}=nothing,
    initial::Union{AbstractVector{<:Real}, Nothing}=nothing,
    )

basis – basis for reconstruction

omegas – array of matrices that provide information about basis functions

method – constant selection method, possible options: "EmpiricalBayes" and "User"

alphas – array of constants, in case method="User" should be provided by user

Fields

basis::Basis

solver::MCMCMatrixUnfolder

source

Reconstruction

Main.StatReg.solve — Function.

solve(
    mcmcunfolder::MCMCUnfolder,
    kernel::Union{Function, AbstractMatrix{<:Real}},
    data::Union{Function, AbstractVector{<:Real}},
    data_errors::Union{Function, AbstractVector{<:Real}},
    y::Union{AbstractVector{<:Real}, Nothing}=nothing;
    model::Union{Model, String} = "Gaussian",
    samples::Int = 10 * 1000,
    burnin::Int = 0,
    thin::Int = 1,
    chains::Int = 1,
    verbose::Bool = false
    )

Arguments

unfolder – model

kernel – discrete or continuous kernel

data – function values

data_errors – function errors

y – points to calculate function values and its errors (when data is given as a function)

model – errors model, "Gaussian" or predefined Mamba.jl model

burnin– numer of initial draws to discard as a burn-in sequence to allow for convergence

thin – step-size between draws to output

chains– number of simulation runs to perform

verbose – whether to print sampler progress at the console

Returns: parameters for mcmc() function.

source

Main.StatReg.solve — Method.

MCMC solver for discrete data and kernel.

solve(
    unfolder::MCMCMatrixUnfolder,
    kernel::AbstractMatrix{<:Real},
    data::AbstractVector{<:Real},
    data_errors::AbstractVecOrMat{<:Real};
    model::Union{Model, String} = "Gaussian",
    samples::Int = 10 * 1000,
    burnin::Int = 0,
    thin::Int = 1,
    chains::Int = 1,
    verbose::Bool = false
    )

Arguments

unfolder – model

kernel – discrete kernel

data – function valuess

data_errors – function errors

model – errors model, "Gaussian" or predefined Mamba.jl model

burnin– numer of initial draws to discard as a burn-in sequence to allow for convergence

thin – step-size between draws to output

chains– number of simulation runs to perform

verbose – whether to print sampler progress at the console

Returns: parameters for mcmc() function.

source

Result

Main.StatReg.get_values — Function.

Allowers to get coefficients and errors from generated data set.

get_values(sim::ModelChains)

Arguments

sim – data generated by mcmc()

Returns: Dict{String, AbstractVector{Real}} with coefficients ("coeff") and errors ("errors").

source

Main.StatReg.PhiVec — Type.

Constructs solution function by coefficients, basis and errors.

PhiVec(coeff::Array{<:Real}, basis::Basis, errors::Array{<:Real})
PhiVec(coeff::Array{<:Real}, basis::Basis)
PhiVec(result::Dict{String, Array{<:Real}}, basis::Basis)

Fields

coeff::Array{<:Real} – coefficients of decomposition of a function in basis

basis::Basis – basis

errors::Union{Array{<:Real}, Nothing} – coefficients of decomposition of a function errors in basis

phi_function(x::Real)::Function – returns constructed function's value at given point

error_function(x::Real)::Union{Function, Nothing} – returns constructed function's error at given point, if errors are specified, otherwise is nothing

source