User's Guide

User's Guide

Kernel

Kernel can be specified as a matrix or as a function. Initialize a kernel as a function:

getOpticsKernels — Method.

getOpticsKernels(name::String, alpha::Float64 = 1.)

Arguments

name - name of a kernel

alpha - kernel function parameter

Returns: kernel, function of 2 variables.

source

Available kernels:

rectangular:

\[K(x, y) = \begin{cases} 1, \text{if } \frac{|x-y|}{\alpha} < 1 \\ 0, \text{otherwise} \end{cases}\]

diffraction:

\[K(x, y) = \left(\frac{sin(\frac{\pi (x-y)}{s_0})}{\frac{\pi (x-y)}{s_0}}\right)^2\]

\[s_0 = \frac{\alpha}{0.886}\]

gaussian:

\[K(x, y) = \frac{2}{\alpha}\sqrt{\frac{\ln2}{\pi}}e^{4\ln2\left(\frac{x-y}{\alpha}\right)^2}\]

triangular:

\[K(x, y) = \begin{cases} \frac{1 - \frac{|x-y|}{\alpha}}{\alpha}, \text{if } \frac{|x-y|}{\alpha} < 1 \\ 0, \text{otherwise} \end{cases}\]

dispersive:

\[K(x, y) = \frac{\alpha}{2 \pi}\left((x-y)^2 + \left(\frac{\alpha}{2}\right)^2\right)\]

exponential:

\[K(x, y) = \frac{\ln2}{\alpha}e^{2\ln2\frac{|x-y|}{\alpha}}\]

heaviside:

\[K(x, y) = \begin{cases} 1, \text{if } x>0 \\ 0, \text{otherwise} \end{cases}\]

discretize_kernel — Method.

discretize_kernel(basis::Basis, kernel::Function, data_points::Array{Float64, 1})

Arguments

basis – basis

kernel – $K(x, y)$, kernel

data_points – array of data points

Returns: discretized kernel K::Array{Float64, 2}, $K_{mn} = \left(\int_a^b K(x, y) \psi_m(x) dx \right)(y_n)$ - matrix of size n$\times$m, where m - number of basis functions, n - number of data points.

source

Basis

Basis — Type.

Abstract type for all bases.

source

BaseFunction — Type.

Type for function with its support.

Constructor

BaseFunction(f, support::Tuple{Float64,Float64})
BaseFunction(f, a::Float64, b::Float64)

Fields

f – function (type depends on the basis)

support::Tuple{Float64,Float64} – support of the function

source

omega — Method.

omega(basis::Basis, order::Int64)

Arguments

basis - basis

order - order of derivatives

Returns: Omega::Array{Float64, 2}, $\Omega_{mn} = \int_a^b \frac{d^{order} \psi_m}{dx^{order}} \frac{d^{order} \psi_n}{dx^{order}}$ - matrix of size n$\times$n of the mean values of derivatives of order order, where n - number of functions in basis.

source

Fourier basis

FourierBasis — Type.

Fourier basis with length $2n+1$: {$0.5$, $\sin(\frac{\pi (x - \frac{a+b}{2})}{b-a})$, $\cos(\frac{\pi (x - \frac{a+b}{2})}{b-a})$, ..., $\sin(\frac{\pi n (x - \frac{a+b}{2})}{b-a})$, $\cos(\frac{\pi n (x - \frac{a+b}{2})}{b-a})$}.

Constructor

FourierBasis(a::Float64, b::Float64, n::Int64)

a, b – the beginning and the end of the segment

n – number of basis functions

Fields

a::Float64 – beginning of the support

b::Float64 – end of the support

n::Int64 – number of basis functions

basis_functions::Array{BaseFunction, 1} – array of basis functions

source

Cubic Spline basis

CubicSplineBasis — Type.

Cubic spline basis - B-spline of the order 3 on given knots with length $n-4$, where n – length of knots array.

Constructor

CubicSplineBasis(
    knots::Array{Float64},
    boundary_condition::Union{Tuple{Union{String, Nothing}, Union{String, Nothing}}, Nothing, String}=nothing
    )

knots – knots of spline

boundary_condition – boundary conditions of basis functions. If tuple, the first element affects left bound, the second element affects right bound. If string, both sides are affected. Possible options: "dirichlet", nothing

Fields

a::Float64 – beginning of the support, matches the first element of the array knots

b::Float64 – end of the support, matches the last element of the array knots

knots::Array{Float64, 1} – array of points on which the spline is built

basis_functions::Array{BaseFunction, 1} – array of basis functions

source

Legendre polynomial basis

LegendreBasis — Type.

Legendre polynomials basis with length $n$.

Constructor

LegendreBasis(a::Float64, b::Float64, n::Int64)

a, b – the beginning and the end of the support

n – number of basis functions

Fields

a::Float64 – beginning of the support

b::Float64 – end of the support

basis_functions::Array{BaseFunction, 1} – array of basis functions

source

Bernstein polynomial basis

BernsteinBasis — Type.

Bernstein polynomials basis.

Constructor

BernsteinBasis(
    a::Float64, b::Float64, n::Int64,
    boundary_condition::Union{Tuple{Union{String, Nothing}, Union{String, Nothing}}, Nothing, String}=nothing
    )

a, b – the beginning and the end of the segment

n – number of basis functions

boundary_condition – boundary conditions of basis functions. If tuple, the first element affects left bound, the second element affects right bound. If string, both sides are affected. Possible options: "dirichlet", nothing.

Fields

a::Float64 – beginning of the support

b::Float64 – end of the support

basis_functions::Array{BaseFunction, 1} – array of basis functions

boundary_condition::Tuple{Union{String, Nothing}, Union{String, Nothing}} – boundary conditions of basis functions. If tuple, the first element affects left bound, the second element affects right bound. If string, both sides are affected. Possible options: "dirichlet", nothing.

source

Gaussian noise distribution with alpha as argmax of posterior probability

Model

GaussErrorMatrixUnfolder — Type.

Model for dicsrete data and kernel.

Constructor

GaussErrorMatrixUnfolder(
    omegas::Array{Array{Float64, 2} ,1},
    method::String="EmpiricalBayes",
    alphas::Union{Array{Float64, 1}, Nothing}=nothing,
    )

omegas – array of matrices that provide information about basis functions

method – constant selection method, possible options: "EmpiricalBayes" and "User"

alphas – array of constants, in case method="User" should be provided by user

low – lower limits for alphas

high – higher limits for alphas

alpha0 – unitial values for alphas

Fields

omegas::Array{Array{Float64, 2} ,1}

n::Int64 – size of square omega matrix

method::String

alphas::Union{Array{Float64}, Nothing}

low::Union{Array{Float64, 1}, Nothing}

high::Union{Array{Float64, 1}, Nothing}

alpha0::Union{Array{Float64, 1}, Nothing}

source

GaussErrorUnfolder — Type.

Model for continuous kernel. Data can be either discrete or continuous.

Constructor

GaussErrorUnfolder(
    basis::Basis,
    omegas::Array{Array{Float64, 2}, 1},
    method::String="EmpiricalBayes",
    alphas::Union{Array{Float64, 1}, Nothing}=nothing,
    low::Union{Array{Float64, 1}, Nothing}=nothing,
    high::Union{Array{Float64, 1}, Nothing}=nothing,
    alpha0::Union{Array{Float64, 1}, Nothing}=nothing
    )

basis – basis for reconstruction

omegas – array of matrices that provide information about basis functions

method – constant selection method, possible options: "EmpiricalBayes" and "User"

alphas – array of constants, in case method="User" should be provided by user

low – lower limits for alphas

high – higher limits for alphas

alpha0 – unitial values for alphas

Fields

basis::Basis

solver::GaussErrorMatrixUnfolder

source

Reconstruction

solve — Method.

solve(
    unfolder::GaussErrorMatrixUnfolder,
    kernel::Array{Float64, 2},
    data::Array{Float64, 1},
    data_errors::Union{Array{Float64, 1}, Array{Float64, 2}},
    )

Arguments

unfolder::GaussErrorMatrixUnfolder – model

kernel::Array{Float64, 2} – discrete kernel

data::Array{Float64, 1} – function values

data_errors::Union{Array{Float64, 1}, Array{Float64, 2}} – function errors

Returns: Dict{String, Array{Float64, 1}} with coefficients ("coeff"), errors ("errors") and optimal constants ("alphas").

source

solve — Method.

solve(
    gausserrorunfolder::GaussErrorUnfolder,
    kernel::Union{Function, Array{Float64, 2}},
    data::Union{Function, Array{Float64, 1}},
    data_errors::Union{Function, Array{Float64, 1}},
    y::Union{Array{Float64, 1}, Nothing},
    )

Arguments

gausserrorunfolder::GaussErrorUnfolder – model

kernel::Union{Function, Array{Float64, 2}} – discrete or continuous kernel

data::Union{Function, Array{Float64, 1}} – function values

data_errors::Union{Function, Array{Float64, 1}} – function errors

y::Union{Array{Float64, 1}, Nothing} – points to calculate function values and its errors (when data is given as a function)

Returns: Dict{String, Array{Float64, 1}} with coefficients ("coeff"), errors ("errors") and optimal constants ("alphas").

source

Any othes noise distribution with alpha as argmax of posterior probability

Model

MCMCMatrixUnfolder — Type.

MCMC model for dicsrete data and kernel.

Constructor

MCMCMatrixUnfolder(
    omegas::Array{Array{Float64, 2} ,1},
    method::String="EmpiricalBayes",
    alphas::Union{Array{Float64, 1}, Nothing}=nothing,
    low::Union{Array{Float64, 1}, Nothing}=nothing,
    high::Union{Array{Float64, 1}, Nothing}=nothing,
    alpha0::Union{Array{Float64, 1}, Nothing}=nothing
    )

omegas – array of matrices that provide information about basis functions

method – constant selection method, possible options: "EmpiricalBayes" and "User"

alphas – array of constants, in case method="User" should be provided by user

low – lower limits for alphas

high – higher limits for alphas

alpha0 – unitial values for alphas

Fields

omegas::Array{Array{Float64, 2} ,1}

n::Int64 – size of square omega matrix

method::String

alphas::Union{Array{Float64}, Nothing}

low::Union{Array{Float64, 1}, Nothing}

high::Union{Array{Float64, 1}, Nothing}

alpha0::Union{Array{Float64, 1}, Nothing}

source

MCMCUnfolder — Type.

MCMC model for continuous kernel. Data can be either discrete or continuous.

Constructor

MCMCUnfolder(
    basis::Basis,
    omegas::Array{Array{Float64, 2}, 1},
    method::String="EmpiricalBayes",
    alphas::Union{Array{Float64, 1}, Nothing}=nothing,
    low::Union{Array{Float64, 1}=nothing,
    high::Union{Array{Float64, 1}, Nothing}=nothing,
    alpha0::Union{Array{Float64, 1}, Nothing}=nothing,
    )

basis – basis for reconstruction

omegas – array of matrices that provide information about basis functions

method – constant selection method, possible options: "EmpiricalBayes" and "User"

alphas – array of constants, in case method="User" should be provided by user

Fields

basis::Basis

solver::MCMCMatrixUnfolder

source

Reconstruction

solve — Function.

solve(
    mcmcunfolder::MCMCUnfolder,
    kernel::Union{Function, Array{Float64, 2}},
    data::Union{Function, Array{Float64, 1}},
    data_errors::Union{Function, Array{Float64, 1}},
    y::Union{Array{Float64, 1}, Nothing}=nothing,
    chains::Int64 = 1,
    samples::Int64 = 10 * 1000
    )

Arguments

gausserrorunfolder::GaussErrorUnfolder – model

kernel::Union{Function, Array{Float64, 2}} – discrete or continuous kernel

data::Union{Function, Array{Float64, 1}} – function values

data_errors::Union{Function, Array{Float64, 1}} – function errors

y::Union{Array{Float64, 1}, Nothing} – points to calculate function values and its errors (when data is given as a function)

chains – number of chains for MCMC integration

samples – number of samples for MCMC integration

Returns: Dict{String, Array{Float64, 1}} with coefficients ("coeff") and errors ("errors").

source

solve — Function.

MCMC solver for dicsrete data and kernel.

solve(
    unfolder::MCMCMatrixUnfolder,
    kernel::Array{Float64, 2},
    data::Array{Float64, 1},
    data_errors::Union{Array{Float64, 1}, Array{Float64, 2}},
    chains::Int64 = 1,
    samples::Int64 = 10 * 1000
    )

Arguments

unfolder::MCMCMatrixUnfolder – model

kernel::Array{Float64, 2} – discrete kernel

data::Array{Float64, 1} – function values

data_errors::Union{Array{Float64, 1}, Array{Float64, 2}} – function errors

chains – number of chains for MCMC integration

samples – number of samples for MCMC integration

Returns: Dict{String, Array{Float64, 1}} with coefficients ("coeff") and errors ("errors").

source

Result

get_values — Method.

Allows to get coefficients and errors from generated data set.

get_values(sim::ModelChains, chains::Int64, n::Int64)

Arguments

sim::ModelChains – data generated by mcmc()

chains::Int64 – number of chains

n::Int64 – variable list length

Returns: Dict{String, Array{Float64, 1}} with coefficients ("coeff") and errors ("errors").

source

PhiVec — Type.

Constructs solution function by coefficients, basis and errors.

Constructor

PhiVec(coeff::Array{Float64}, basis::Basis, errors::Array{Float64})
PhiVec(coeff::Array{Float64}, basis::Basis)
PhiVec(result::Dict{String, Array{Float64}}, basis::Basis)

Fields

coeff::Array{Float64} – coefficients of decomposition of a function in basis

basis::Basis – basis

errors::Union{Array{Float64}, Nothing} – coefficients of decomposition of a function errors in basis

phi_function(x::Float64)::Function – returns constructed function's value at given point

error_function(x::Float64)::Union{Function, Nothing} – returns constructed function's error at given point,

if errors are specified, otherwise is nothing

source